pyproject.toml
[tool.poetry]
name = "etf-volatility-harvest"
version = "0.1.0"
description = ""
authors = ["Moses McCall <mosesmc52@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.2.1"
xtermcolor = "^1.3"
boto3 = "^1.42.44"
pandas = "^3.0.0"
alpaca-py = "^0.43.2"
pandas-market-calendars = "^5.3.0"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.5.3"
yfinance = "^1.1.0"
seaborn = "^0.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
